Split script.cpp to smaller files in src/script
