+++ src/Makefile.qt.include
@@ -263,7 +263,7 @@ BITCOIN_QT_CPP = \
   rpcwallet.cpp \
   rpcblockchain.cpp \
   rpcrawtransaction.cpp \
-  script.cpp \
+  script/script.cpp \
   sync.cpp \
   util.cpp \
   wallet.cpp \
