+++ src/test/sigopcount_tests.cpp
@@ -2,7 +2,7 @@
 #include <boost/test/unit_test.hpp>
 #include <boost/foreach.hpp>
 
-#include "script.h"
+#include "script/script.h"
 #include "key.h"
 
 using namespace std;
